modify label.rb file
