+++ classes/label.rb
@@ -1,8 +1,8 @@
 class Label
     attr_reader :id, :items, :title, :color
   
-  def initialize(title, color)
-    @id = Random.rand(1000)
+    def initialize(title, color, id = Random.rand(10_000))
+    @id = id
     @items = []
     @title = title
     @color = color
@@ -13,11 +13,4 @@ class Label
     item.label = self
   end
 
-  def create_json
-    {
-       id: @id,
-       title: @title,
-       color: @color
-    }
-    end  
 end
